refactor: build using rollup.js
