+++ .npmignore
@@ -3,3 +3,4 @@ test
 coverage
 .*
 *.log
+rollup.config.js
