+++ package.json
@@ -12,12 +12,12 @@
   },
   "description": "Formats data into a string table.",
   "devDependencies": {
-    "@babel/cli": "^7.10.5",
     "@babel/core": "^7.11.4",
-    "@babel/node": "^7.10.5",
-    "@babel/plugin-transform-flow-strip-types": "^7.10.4",
     "@babel/preset-env": "^7.11.0",
     "@babel/register": "^7.10.5",
+    "@rollup/plugin-babel": "^5.2.1",
+    "@rollup/plugin-commonjs": "^16.0.0",
+    "@rollup/plugin-node-resolve": "^10.0.0",
     "ajv-cli": "^3.2.1",
     "ajv-keywords": "^3.5.2",
     "babel-plugin-istanbul": "^6.0.0",
@@ -29,11 +29,12 @@
     "eslint": "^7.7.0",
     "eslint-config-canonical": "^23.0.1",
     "flow-bin": "^0.132.0",
-    "flow-copy-source": "^2.0.9",
     "gitdown": "^3.1.3",
     "husky": "^4.2.5",
+    "mkdirp": "^1.0.4",
     "mocha": "^8.1.3",
     "nyc": "^15.1.0",
+    "rollup": "^2.33.3",
     "semantic-release": "^17.1.1",
     "sinon": "^9.0.3"
   },
@@ -74,10 +75,10 @@
     "url": "https://github.com/gajus/table"
   },
   "scripts": {
-    "build": "rm -fr ./dist && NODE_ENV=production babel ./src --out-dir ./dist --copy-files --source-maps && npm run create-validators && flow-copy-source src dist",
+    "build": "rm -fr ./dist && mkdirp ./dist && npm run create-validators && rollup -c",
     "create-readme": "gitdown ./.README/README.md --output-file ./README.md",
-    "create-validators": "ajv compile --all-errors --inline-refs=false -s src/schemas/config -c ajv-keywords/keywords/typeof -o dist/validateConfig.js && ajv compile --all-errors --inline-refs=false -s src/schemas/streamConfig -c ajv-keywords/keywords/typeof -o dist/validateStreamConfig.js",
-    "lint": "npm run build && eslint ./src ./test && flow",
+    "create-validators": "rm -fr ./src/compiled_schema && mkdirp ./src/compiled_schema && ajv compile --all-errors --inline-refs=false -s src/schemas/config -c ajv-keywords/keywords/typeof -o src/compiled_schema/validateConfig.js && ajv compile --all-errors --inline-refs=false -s src/schemas/streamConfig -c ajv-keywords/keywords/typeof -o src/compiled_schema/validateStreamConfig.js",
+    "lint": "npm run build && eslint --ignore-path .gitignore && flow",
     "test": "mocha --require @babel/register"
   },
   "version": "1.0.0"
